+++ lib/generator.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/dart/element/type.dart';
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
@@ -10,7 +9,7 @@ class BuiltReduxGenerator extends Generator {
     final result = new StringBuffer();
     var hasWrittenHeaders = false;
     for (final element in library.allElements) {
-      if (_needsReduxActions(element) && element is ClassElement) {
+      if (_isReduxActions(element) && element is ClassElement) {
         if (!hasWrittenHeaders) {
           hasWrittenHeaders = true;
           result.writeln(_lintIgnores);
@@ -27,14 +26,81 @@ class BuiltReduxGenerator extends Generator {
 const _lintIgnores = """
 // ignore_for_file: avoid_classes_with_only_static_members
 // ignore_for_file: annotate_overrides
+// ignore_for_file: overridden_fields
 """;
 
-String _generateActions(ClassElement element) =>
-    _generateDispatchersIfNeeded(element) + _actionNamesClassTemplate(element);
+ActionsClass _actionsClassFromElement(ClassElement element) => new ActionsClass(
+      element.name,
+      _actionsFromElement(element).toSet(),
+      _composedActionClasses(element).toSet(),
+      _actionsClassFromInheritedElements(element).toSet(),
+    );
 
-String _generateDispatchersIfNeeded(ClassElement element) =>
+Iterable<ComposedActionClass> _composedActionClasses(ClassElement element) =>
+    element.fields
+        .where((f) => _isReduxActions(f.type.element))
+        .map((f) => new ComposedActionClass(f.name, f.type.name));
+
+Iterable<Action> _actionsFromElement(ClassElement element) => element.fields
+    .where(_isActionDispatcher)
+    .map((field) => _fieldElementToAction(element, field));
+
+Iterable<ActionsClass> _actionsClassFromInheritedElements(
+        ClassElement element) =>
+    element.allSupertypes
+        .map((s) => s.element)
+        .where(_isReduxActions)
+        .map(_actionsClassFromElement);
+
+Action _fieldElementToAction(ClassElement element, FieldElement field) =>
+    new Action('${element.name}-${field.name}', field.name,
+        _fieldType(element, field));
+
+// hack to return the generics for the action
+// this is used so action whose payloads are of generated types
+// will not result in dynamic
+String _fieldType(ClassElement element, FieldElement field) {
+  if (field.isSynthetic) {
+    return _syntheticFieldType(element, field);
+  }
+  return _getGenerics(field.source.contents.data, field.nameOffset);
+}
+
+String _syntheticFieldType(ClassElement element, FieldElement field) {
+  final method = element.getGetter(field.name);
+  return _getGenerics(method.source.contents.data, method.nameOffset);
+}
+
+String _getGenerics(String source, int nameOffset) {
+  final trimAfterName = source.substring(0, nameOffset);
+  final trimBeforeActionDispatcher =
+      trimAfterName.substring(trimAfterName.lastIndexOf('ActionDispatcher'));
+  return trimBeforeActionDispatcher.substring(
+      trimBeforeActionDispatcher.indexOf('<') + 1,
+      trimBeforeActionDispatcher.lastIndexOf('>'));
+}
+
+bool _isReduxActions(Element element) =>
+    element is ClassElement && _hasSuperType(element, 'ReduxActions');
+
+bool _isActionDispatcher(FieldElement element) =>
+    element.type.name == 'ActionDispatcher';
+
+bool _hasSuperType(ClassElement classElement, String type) =>
+    classElement.allSupertypes
+        .any((interfaceType) => interfaceType.name == type) &&
+    !classElement.displayName.startsWith('_\$');
+
+String _generateActions(ClassElement element) {
+  final actionClass = _actionsClassFromElement(element);
+  return _generateDispatchersIfNeeded(element, actionClass) +
+      _actionNamesClassTemplate(actionClass);
+}
+
+String _generateDispatchersIfNeeded(
+        ClassElement element, ActionsClass actionsClass) =>
     element.constructors.length > 1
-        ? _actionDispatcherClassTemplate(element)
+        ? _actionDispatcherClassTemplate(actionsClass)
         : '';
 
 /*
@@ -43,167 +109,120 @@ String _generateDispatchersIfNeeded(ClassElement element) =>
 
 */
 
-String _actionDispatcherClassTemplate(ClassElement element) => '''
-  class _\$${element.name} extends ${element.name}{
-    factory _\$${element.name}() => new _\$${element.name}._();
-    _\$${element.name}._() : super._();
+String _actionDispatcherClassTemplate(ActionsClass actionsClass) => '''
+  class _\$${actionsClass.className} extends ${actionsClass.className}{
+    factory _\$${actionsClass.className}() => new _\$${actionsClass.className}._();
+    _\$${actionsClass.className}._() : super._();
 
-    ${_allActionDispatcherFieldsTemplate(element)}
+    ${_allActionDispatcherFieldsTemplate(actionsClass)}
+    ${_allComposedActionClassesFieldsTemplate(actionsClass)}
 
     @override
     void setDispatcher(Dispatcher dispatcher) {
-      ${_allSetDispatchersTemplate(element)}
+      ${_allActionDispatcherSetDispatchersTemplate(actionsClass)}
+      ${_allComposedActionClassesSetDispatchersTemplate(actionsClass)}
     }
   }
 ''';
 
-String _allActionDispatcherFieldsTemplate(ClassElement element) =>
-    _forEachFieldWithInherited(
-        element, _actionDispatcherFieldTemplate, _isActionDispatcher) +
-    _forEachFieldWithInherited(
-        element, _reduxActionFieldTemplate, _needsReduxActions);
+String _allActionDispatcherFieldsTemplate(ActionsClass actionsClass) =>
+    actionsClass.allActions.fold(
+        '', (comb, next) => '$comb\n${_actionDispatcherFieldTemplate(next)}');
 
-String _actionDispatcherFieldTemplate(ClassElement e, FieldElement f) {
-  final genericType =
-      _getActionGenericType((f.type as InterfaceType).typeArguments.first);
-  return 'final ActionDispatcher<${genericType}> ${f.name} = new ActionDispatcher<${genericType}>(\'${e.name}-${f.name}\');';
-}
+String _allComposedActionClassesFieldsTemplate(ActionsClass actionsClass) =>
+    actionsClass.allComposed.fold('',
+        (comb, next) => '$comb\n${_composedActionClassesFieldTemplate(next)}');
 
-String _reduxActionFieldTemplate(ClassElement e, FieldElement f) {
-  final typeName = f.type.element.name;
-  return 'final $typeName ${f.name} = new $typeName();';
-}
+String _actionDispatcherFieldTemplate(Action action) =>
+    'final ${action.fieldName} = new  ActionDispatcher<${action.type}>(\'${action.actionName}\');';
 
-String _allSetDispatchersTemplate(ClassElement element) =>
-    _forEachFieldWithInherited(
-        element, _setDispatcheTemplate, _isActionDispatcher) +
-    _forEachFieldWithInherited(
-        element, _setDispatcheTemplate, _needsReduxActions);
+String _composedActionClassesFieldTemplate(
+        ComposedActionClass composedActionClass) =>
+    'final ${composedActionClass.fieldName} = new ${composedActionClass.type}();';
 
-String _setDispatcheTemplate(ClassElement e, FieldElement f) =>
-    '${f.name}.setDispatcher(dispatcher);';
+String _allActionDispatcherSetDispatchersTemplate(ActionsClass actionsClass) =>
+    actionsClass.allActions.fold(
+        '', (comb, next) => '$comb\n${_setDispatcheTemplate(next.fieldName)}');
 
-/*
+String _allComposedActionClassesSetDispatchersTemplate(
+        ActionsClass actionsClass) =>
+    actionsClass.allComposed.fold(
+        '', (comb, next) => '$comb\n${_setDispatcheTemplate(next.fieldName)}');
 
-  Action Names
+String _setDispatcheTemplate(String fieldName) =>
+    '${fieldName}.setDispatcher(dispatcher);';
 
-*/
+// /*
 
-String _actionNamesClassTemplate(ClassElement element) => '''
-  class ${element.name}Names {
-    ${_allActionNamesFieldsTemplate(element)}
+//   Action Names
+
+// */
+
+String _actionNamesClassTemplate(ActionsClass actionsClass) => '''
+  class ${actionsClass.className}Names {
+    ${_allActionNamesFieldsTemplate(actionsClass)}
   }
 ''';
 
-String _allActionNamesFieldsTemplate(ClassElement element) =>
-    _forEachFieldWithInherited(
-        element, _actionNameTemplate, _isActionDispatcher);
+String _allActionNamesFieldsTemplate(ActionsClass actionsClass) =>
+    actionsClass.allActions
+        .fold('', (comb, next) => '$comb\n${_actionNameTemplate(next)}');
 
-String _actionNameTemplate(ClassElement e, FieldElement f) {
-  final genericType =
-      _getActionGenericType((f.type as InterfaceType).typeArguments.first);
-  return 'static final ActionName<${genericType}> ${f.name} = new ActionName<${genericType}>(\'${e.name}-${f.name}\');';
+String _actionNameTemplate(Action action) =>
+    'static final ${action.fieldName} = new ActionName<${action.type}>(\'${action.actionName}\');';
+
+class ActionsClass {
+  final String className;
+  final Set<Action> actions;
+  final Set<ComposedActionClass> composed;
+  final Set<ActionsClass> inherited;
+  ActionsClass(this.className, this.actions, this.composed, this.inherited);
+  Set<Action> get allActions => new Set<Action>.from(
+        actions.toList()
+          ..addAll(inherited.map((ac) => ac.actions).expand((a) => a)),
+      );
+  Set<ComposedActionClass> get allComposed => new Set<ComposedActionClass>.from(
+        composed.toList()
+          ..addAll(inherited.map((ac) => ac.composed).expand((c) => c)),
+      );
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) return true;
+    return other is ActionsClass && className == other.className;
+  }
+
+  @override
+  int get hashCode => className.hashCode;
 }
 
-/*
+class Action {
+  final String actionName;
+  final String fieldName;
+  final String type;
+  Action(this.actionName, this.fieldName, this.type);
 
-  Util
-
-*/
-String _forEachField(
-  ClassElement element,
-  String fieldTemplate(ClassElement e, FieldElement f),
-  bool fieldTest(Element e),
-) =>
-    element.fields
-        .where(
-          (f) => fieldTest(f.type.element),
-        )
-        .fold(
-          '',
-          (combined, next) => '$combined${fieldTemplate(element, next)}',
-        );
-
-String _forEachFieldWithInherited(
-  ClassElement element,
-  String fieldTemplate(ClassElement e, FieldElement f),
-  bool fieldTest(Element e),
-) =>
-    element.allSupertypes
-        .map(
-          (s) => _forEachField(s.element, fieldTemplate, fieldTest),
-        )
-        .fold(
-          _forEachField(element, fieldTemplate, fieldTest),
-          (combined, next) => '$combined$next',
-        );
-
-bool _needsReduxActions(Element element) =>
-    element is ClassElement && _hasSuperType(element, 'ReduxActions');
-
-bool _isActionDispatcher(Element element) =>
-    element is ClassElement && element.name == 'ActionDispatcher';
-
-bool _hasSuperType(ClassElement classElement, String type) =>
-    classElement.allSupertypes
-        .any((interfaceType) => interfaceType.name == type) &&
-    !classElement.displayName.startsWith('_\$');
-
-String _getActionGenericType(DartType type) {
-  if (type is FunctionType) {
-    final generics =
-        _correctUnresolvedGenerics(type.typeArguments, type.typeParameters);
-    if (generics.isEmpty) return typeNameOf(type);
-    return typeNameOf(type) + '<${generics.join(',')}>';
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) return true;
+    return other is Action && fieldName == other.fieldName;
   }
 
-  if (type is ParameterizedType) {
-    final generics =
-        _correctUnresolvedGenerics(type.typeArguments, type.typeParameters);
-    if (generics.isEmpty) return type.name;
-    return type.name + '<${generics.join(',')}>';
-  }
-
-  if (type.isVoid) return 'void';
-
-  return 'dynamic';
+  @override
+  int get hashCode => fieldName.hashCode;
 }
 
-List<String> _correctUnresolvedGenerics(
-    List<DartType> args, List<TypeParameterElement> params) {
-  // hack for thunks/repatches or any other type argument list where
-  // any given argument is a Built and the proceding is a Builder
-  // and the builder is dynamic in the typeArguments list becauses it is
-  // yet to be generated. This is complex and a bit awkward but
-  // it is written this way to be very careful not to make any unintended
-  // changes the the typeArguments list.
-  final typeArguments = args.map((ta) => ta.toString()).toList();
-  final boundParams = params.map((e) => e.bound);
-  for (int i = 0; i < boundParams.length; i++) {
-    // get the bound param at this spot
-    final b = boundParams.elementAt(i);
+class ComposedActionClass {
+  final String fieldName;
+  final String type;
+  ComposedActionClass(this.fieldName, this.type);
 
-    // if it is a Built and there is another argument after it
-    // check if it is supposed to be a Builder
-    if (b != null &&
-        b.name.startsWith('Built') &&
-        i != boundParams.length - 1) {
-      final next = boundParams.elementAt(i + 1);
-      if (next != null &&
-          next.name.startsWith('Builder') &&
-          typeArguments.elementAt(i + 1) == 'dynamic') {
-        log.info(
-            "replacing unresolved builder name ${typeArguments.elementAt(i)}Builder");
-        // if it is a builder replace typeArguments at this location as
-        // the Built's name + Builder
-        typeArguments.replaceRange(
-          i + 1,
-          i + 2,
-          ['${typeArguments.elementAt(i)}Builder'],
-        );
-      }
-    }
+  @override
+  bool operator ==(Object other) {
+    if (identical(other, this)) return true;
+    return other is Action && fieldName == other.fieldName;
   }
 
-  return typeArguments;
+  @override
+  int get hashCode => fieldName.hashCode;
 }