+++ test/unit/nested_models.g.dart
@@ -8,52 +8,53 @@ part of nested_models;
 
 // ignore_for_file: avoid_classes_with_only_static_members
 // ignore_for_file: annotate_overrides
+// ignore_for_file: overridden_fields
 
 class _$BaseActions extends BaseActions {
   factory _$BaseActions() => new _$BaseActions._();
   _$BaseActions._() : super._();
 
-  final ActionDispatcher<Null> baseAction =
-      new ActionDispatcher<Null>('BaseActions-baseAction');
-  final ChildActions child = new ChildActions();
+  final baseAction = new ActionDispatcher<Null>('BaseActions-baseAction');
+
+  final child = new ChildActions();
 
   @override
   void setDispatcher(Dispatcher dispatcher) {
     baseAction.setDispatcher(dispatcher);
+
     child.setDispatcher(dispatcher);
   }
 }
 
 class BaseActionsNames {
-  static final ActionName<Null> baseAction =
-      new ActionName<Null>('BaseActions-baseAction');
+  static final baseAction = new ActionName<Null>('BaseActions-baseAction');
 }
 
 class _$ChildActions extends ChildActions {
   factory _$ChildActions() => new _$ChildActions._();
   _$ChildActions._() : super._();
 
-  final ActionDispatcher<Null> childAction =
-      new ActionDispatcher<Null>('ChildActions-childAction');
-  final GrandchildActions grandchild = new GrandchildActions();
+  final childAction = new ActionDispatcher<Null>('ChildActions-childAction');
+
+  final grandchild = new GrandchildActions();
 
   @override
   void setDispatcher(Dispatcher dispatcher) {
     childAction.setDispatcher(dispatcher);
+
     grandchild.setDispatcher(dispatcher);
   }
 }
 
 class ChildActionsNames {
-  static final ActionName<Null> childAction =
-      new ActionName<Null>('ChildActions-childAction');
+  static final childAction = new ActionName<Null>('ChildActions-childAction');
 }
 
 class _$GrandchildActions extends GrandchildActions {
   factory _$GrandchildActions() => new _$GrandchildActions._();
   _$GrandchildActions._() : super._();
 
-  final ActionDispatcher<Null> grandchildAction =
+  final grandchildAction =
       new ActionDispatcher<Null>('GrandchildActions-grandchildAction');
 
   @override
@@ -63,7 +64,7 @@ class _$GrandchildActions extends GrandchildActions {
 }
 
 class GrandchildActionsNames {
-  static final ActionName<Null> grandchildAction =
+  static final grandchildAction =
       new ActionName<Null>('GrandchildActions-grandchildAction');
 }